add Professor view: -change faculty -get schedule -change username; add Student view: -change faculty -get schedule -change username; rewrite tests and add more unhappy path.

+++ src/main/java/ua/foxminded/pskn/universitycms/repository/university/ScheduleRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 @Repository
 public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
-    List<Schedule> findScheduleByGroupId(Long groupId);
+    List<Schedule> findScheduleByGroupId(int groupId);
 
     List<Schedule> findScheduleByProfessorId(int professorId);
 }
